+++ src/test/java/restaurant_tests/interactors/RestaurantCreateInteractorTest.java
@@ -1,4 +1,4 @@
-package restaurant_tests;
+package restaurant_tests.interactors;
 
 import entities.OwnerFactory;
 import entities.Restaurant;
@@ -7,7 +7,9 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
 import restaurant_use_case.gateways.RestaurantDSGateway;
-import restaurant_use_case.interactors.*;
+import restaurant_use_case.interactors.FileRestaurant;
+import restaurant_use_case.interactors.RestaurantRequestModel;
+import restaurant_use_case.interactors.CreateRestaurantInteractor;
 import restaurant_use_case.interfaces.RestaurantInputBoundary;
 import restaurant_use_case.screens.RestaurantPresenter;
 import restaurant_use_case.screens.RestaurantResponseModel;
@@ -16,12 +18,13 @@ import user_use_case.interfaces.UserGatewayInterface;
 
 import java.io.IOException;
 
+
 /**
- * Tests the RestaurantEditInteractor for correct response and action performed
+ * Tests the RestaurantCreateInteractor for the correct response and action performed
  */
-public class RestaurantEditInteractorTest {
+public class RestaurantCreateInteractorTest {
     /**
-     * The Restaurant gateway responsible for managing the temporary test database
+     * The testing file that will be written and read from within the Restaurant tests
      */
     final RestaurantDSGateway restaurantGateway = new FileRestaurant("src/test/java/restaurant_tests/temptest.csv");
     /**
@@ -29,63 +32,52 @@ public class RestaurantEditInteractorTest {
      */
     final RestaurantFactory factory = new RestaurantFactory();
     /**
-     * The User gateway
+     * The User database gateway
      */
     final UserGatewayInterface userGateway = new UserGateway();
 
-    public RestaurantEditInteractorTest() throws IOException {
+    public RestaurantCreateInteractorTest() throws IOException {
     }
 
+    /**
+     * Simulates a Restaurant Presenter that receives the information that the Interact creates
+     */
     @Test
     void simulation() {
         // Temporary test presenter
         RestaurantPresenter presenter = new RestaurantPresenter() {
             @Override
             public RestaurantResponseModel prepareSuccessView(RestaurantResponseModel responseModel) {
-                // Get the Restaurant edited by the interactor
-                Restaurant editedRestaurant = responseModel.getRestaurant();
-                // Check that the old values have been changed to the new values
-                assertEquals("newRestaurant", editedRestaurant.getName());
-                assertEquals("fast food", editedRestaurant.getCuisineType());
-                // Check that the old values that have not been changed have been retained
-                assertEquals(3, editedRestaurant.getPriceBucket());
-                // Check that the unique location identifier has not been changed and the restaurant remains in database
-                assertEquals("123456", editedRestaurant.getLocation());
+                // Checks the Restaurant returned by the interactor
+                Restaurant newRestaurant = responseModel.getRestaurant();
+                assertEquals("newRestaurant", newRestaurant.getName());
+                assertEquals("123456", newRestaurant.getLocation());
+                // Check that the restaurant has been successfully stored in the database
                 assertTrue(restaurantGateway.existsByLocation("123456"));
-                // Reset simulation database
                 restaurantGateway.deleteRestaurant("123456");
+                // Checks the operation returned by the interactor
+                // Checks mutability of operation
                 return null;
             }
 
             @Override
             public RestaurantResponseModel prepareFailView(String error) {
-                // Shouldn't reach this branch
+                // Should not reach this branch
                 fail(error);
                 return null;
             }
         };
         // Setup the interactor
-        RestaurantInputBoundary interactor = new EditRestaurantInteractor(restaurantGateway, presenter);
+        RestaurantInputBoundary interactor = new CreateRestaurantInteractor(factory, restaurantGateway, userGateway, presenter);
         OwnerFactory userFactory = new OwnerFactory();
-        // The "old restaurant" we want to edit
-        Restaurant oldRestaurant = factory.create("0000",
-                "oldRestaurant",
-                "123456",
-                "bbq",
-                3);
         // Setup the input data that would normally be done by the controller
-        RestaurantEditRequestModel inputData = new RestaurantEditRequestModel(
+        RestaurantRequestModel inputData = new RestaurantRequestModel(
                 userFactory.CreateUserObject("0000","1234"),
-                // The new information, changed Restaurant Name, cuisine type
                 "newRestaurant",
                 "123456",
-                "fast food",
-                3,
-                oldRestaurant
+                "bbq",
+                3
         );
-        // Make the Restaurant exist
-        restaurantGateway.save(oldRestaurant);
-        // Edit
         interactor.create(inputData);
 
     }
